+++ src/js/react/FriendsUI/Orders.js
@@ -84,9 +84,7 @@ export default function Orders() {
           ))}
         </TableBody>
       </Table>
-      <Link color="primary" href="#" onClick={preventDefault} sx={{ mt: 3 }}>
-        See more orders
-      </Link>
+
     </React.Fragment>
   );
 }
