--- src/js/react/FriendsUI/listItems.js
@@ -1,69 +1,0 @@
-import * as React from 'react';
-import ListItemButton from '@mui/material/ListItemButton';
-import ListItemIcon from '@mui/material/ListItemIcon';
-import ListItemText from '@mui/material/ListItemText';
-import ListSubheader from '@mui/material/ListSubheader';
-import DashboardIcon from '@mui/icons-material/Dashboard';
-import ShoppingCartIcon from '@mui/icons-material/ShoppingCart';
-import PeopleIcon from '@mui/icons-material/People';
-import BarChartIcon from '@mui/icons-material/BarChart';
-import LayersIcon from '@mui/icons-material/Layers';
-import AssignmentIcon from '@mui/icons-material/Assignment';
-import PeopleOutlineIcon from '@mui/icons-material/PeopleOutline';
-import PersonAddAltIcon from '@mui/icons-material/PersonAddAlt';
-
-const styles ={
-  marginBottom:'20%',
-}
-
-
-
-export const mainListItems = (
-  <React.Fragment>
-    <ListItemButton style={styles}>
-      <ListItemIcon>
-        <PeopleIcon />
-      </ListItemIcon>
-      <ListItemText primary="Friends" />
-    </ListItemButton>
-    <ListItemButton style={styles}>
-      <ListItemIcon>
-        <PersonAddAltIcon />
-      </ListItemIcon>
-      <ListItemText primary="Friend Requests" />
-    </ListItemButton>
-    <ListItemButton style={styles}>
-      <ListItemIcon>
-        <PeopleOutlineIcon />
-      </ListItemIcon>
-      <ListItemText primary="Followers" />
-    </ListItemButton>
-
-  </React.Fragment>
-);
-
-export const secondaryListItems = (
-  <React.Fragment>
-    <ListSubheader component="div" inset>
-      Saved reports
-    </ListSubheader>
-    <ListItemButton>
-      <ListItemIcon>
-        <AssignmentIcon />
-      </ListItemIcon>
-      <ListItemText primary="Current month" />
-    </ListItemButton>
-    <ListItemButton>
-      <ListItemIcon>
-        <AssignmentIcon />
-      </ListItemIcon>
-      <ListItemText primary="Last quarter" />
-    </ListItemButton>
-    <ListItemButton>
-      <ListItemIcon>
-        <AssignmentIcon />
-      </ListItemIcon>
-      <ListItemText primary="Year-end sale" />
-    </ListItemButton>
-  </React.Fragment>
-);
